# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Recetas",
        "version": "1.0"
    },
    "paths": {
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Ver todas las categorías",
                "operationId": "684188b8d25abcad5e1c8a786226c02b",
                "responses": {
                    "200": {
                        "description": "Listado de categorías"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Crear una nueva categoría",
                "operationId": "240f085d72cd493a2f869e93edc2c132",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada"
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Ver una categoría específica",
                "operationId": "0644cd38b2452f7e5c9495574afcba4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría encontrada"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Actualizar una categoría",
                "operationId": "8124c319ebe2032b571b923486a66a77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Eliminar una categoría",
                "operationId": "266668bb5896ce68b7848fb39bf8107e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría eliminada"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            }
        },
...
    },
...
}